move subscriptions file
